+++ src/exactly_lib/instructions/utils/parse/__init__.py
@@ -1,3 +1,0 @@
-"""
-Utilities for parsing instruction arguments.
-"""