+++ src/exactly_lib/test_case_utils/string_transformer/impl/run_program/parse.py
@@ -1,9 +1,8 @@
 from typing import Sequence
 
-from exactly_lib.definitions import misc_texts, formatting
+from exactly_lib.definitions import misc_texts
 from exactly_lib.definitions.cross_ref.app_cross_ref import SeeAlsoTarget
 from exactly_lib.definitions.entity import syntax_elements, types
-from exactly_lib.processing import exit_values
 from exactly_lib.section_document.element_parsers.token_stream_parser import TokenParser
 from exactly_lib.symbol.logic.string_transformer import StringTransformerSdv
 from exactly_lib.test_case_utils.documentation import texts
@@ -35,14 +34,10 @@ class SyntaxDescription(grammar.PrimitiveExpressionDescriptionWithNameAsInitialS
     def description_rest(self) -> Sequence[ParagraphItem]:
         tp = TextParser({
             'program': types.PROGRAM_TYPE_INFO.name,
-            'The_program_type_must_terminate': texts.THE_PROGRAM_TYPE_MUST_TERMINATE_SENTENCE,
             'stdin': misc_texts.STDIN,
             'stdout': misc_texts.STDOUT,
-            'exit_code': misc_texts.EXIT_CODE,
-            'HARD_ERROR': exit_values.EXECUTION__HARD_ERROR.exit_identifier,
-            'ignore_exit_code_option': formatting.argument_option(names.RUN_WITH_IGNORED_EXIT_CODE_OPTION_NAME),
         })
-        return tp.fnap(_DESCRIPTION_REST)
+        return tp.fnap(_DESCRIPTION_REST) + texts.run_with_ignored_exit_code_option()
 
     @property
     def see_also_targets(self) -> Sequence[SeeAlsoTarget]:
@@ -56,11 +51,4 @@ _PROGRAM_PARSER = parse_program.program_parser(
 
 _DESCRIPTION_REST = """\
 Runs {program:a}, with input and output via {stdin} and {stdout}.
-
-
-The result is {HARD_ERROR} if the {exit_code} is non-zero,
-unless {ignore_exit_code_option} is given.
-
-
-{The_program_type_must_terminate}
 """