+++ test/exactly_lib_test/instructions/multi_phase/new_file/test_resources/arguments_building.py
@@ -1,7 +1,9 @@
 from exactly_lib.definitions import instruction_arguments
+from exactly_lib.definitions.primitives import program as program_options
 from exactly_lib.definitions.primitives import string_transformer
-from exactly_lib.instructions.utils.parse import parse_file_maker
+from exactly_lib.instructions.utils.file_maker import defs
 from exactly_lib.section_document.parse_source import ParseSource
+from exactly_lib.util.cli_syntax import option_syntax
 from exactly_lib.util.process_execution.process_output_files import ProcOutputFile
 from exactly_lib_test.instructions.test_resources import parse_file_maker as file_maker_args
 from exactly_lib_test.section_document.test_resources.parse_source import remaining_source
@@ -15,13 +17,22 @@ from exactly_lib_test.test_resources.strings import WithToString
 def from_program(file: WithToString,
                  output_variant: ProcOutputFile,
                  program: ArgumentElements,
-                 transformation: WithToString = None) -> ArgumentElements:
-    program_with_pgm_option = program.prepend_to_first_line([
-        ab.option(parse_file_maker.PROGRAM_OUTPUT_OPTIONS[output_variant])
-    ])
+                 transformation: WithToString = None,
+                 ignore_exit_code: bool = False,
+                 ) -> ArgumentElements:
+    first_line = [
+        file,
+        instruction_arguments.ASSIGNMENT_OPERATOR,
+        option_syntax.option_syntax(defs.PROGRAM_OUTPUT_OPTIONS[output_variant]),
+    ]
+    if ignore_exit_code:
+        first_line.append(
+            option_syntax.option_syntax(program_options.WITH_IGNORED_EXIT_CODE_OPTION_NAME)
+        )
 
-    ret_val = ArgumentElements([file, instruction_arguments.ASSIGNMENT_OPERATOR]) \
-        .followed_by(program_with_pgm_option)
+    ret_val = ArgumentElements(first_line) \
+        .followed_by(program)
+
     if transformation is None:
         return ret_val
     else: