+++ test/exactly_lib_test/test_case_utils/program/test_resources/arguments_building.py
@@ -69,6 +69,10 @@ def interpret_py_source_file(py_file: WithToString) -> ArgumentElementsRenderer:
     ])
 
 
+def interpret_py_source_file_elements(py_file: WithToString) -> ArgumentElements:
+    return ArgumentElements([interpret_py_source_file(py_file)])
+
+
 def program(program_arg: WithToString,
             transformation=None) -> ArgumentElements:
     extra = []