+++ CHANGELOG.md
@@ -12,6 +12,7 @@ with exception of a "0." prefix until version 1 is released.
  - Type `file-matcher` - primitive `run`
  - Type `string-matcher` - primitive `run`
  - Type `string-transformer` - primitive `run`
+ - Instruction `file` - `-ignore-exit-code` option
 
 ### Fixed